add : Doc Detail

+++ src/main/java/com/DB/Project/Entitiy/User.java
@@ -5,7 +5,7 @@ import lombok.*;
 @Getter
 @Setter
 public class User {
-    private Long userID;
+    private Integer userID;
     private String name;
     private String id;
     private String passwd;
